cloud/gcp/requirements.txt
# Google Cloud Platform SDK
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
joblib>=1.2.0

# 地理数据处理
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.4.0

# 深度学习（GCP基础镜像已包含TensorFlow 2.15，自带 Keras 2.x）
# 不要再单独装 / 卸载 keras
# tensorflow>=2.15.0  # 已在基础镜像中

# ✅ 这里改成 0.11.0，而不是 0.12.0
scikeras==0.11.0

# 可视化
matplotlib>=3.7.0

# 数据科学工具
xarray>=2023.1.0
dask>=2023.1.0

# 进度条
tqdm>=4.65.0

shap>=0.42.0

# 其他工具
pathlib2>=2.3.7; python_version < '3.6'
